CLEANUP: remove superseded shared-helper interfaces
