+++ mne_denoise/_filtering.py
@@ -8,7 +8,7 @@ several denoisers so callers pass Hz-domain edges directly via
 from __future__ import annotations
 
 import numpy as np
-from scipy.signal import butter, sosfiltfilt
+from scipy.signal import butter
 
 
 def design_butter_sos(
@@ -23,16 +23,3 @@ def design_butter_sos(
     ``fs=sfreq`` rather than by the caller.
     """
     return butter(order, freqs, btype=btype, fs=sfreq, output="sos")
-
-
-def _filter_channels(
-    data: np.ndarray,
-    filter_spec: tuple[str, float | tuple[float, float]] | None,
-    sfreq: float,
-) -> np.ndarray:
-    """Filter along the last axis with a zero-phase 4th-order Butterworth."""
-    if filter_spec is None:
-        return data
-    btype, freqs = filter_spec
-    sos = design_butter_sos(4, freqs, btype, sfreq)
-    return sosfiltfilt(sos, data, axis=-1)
