+++ mne_denoise/dss/variants/tsr.py
@@ -533,7 +533,6 @@ class TimeShiftDSS(BaseEstimator, TransformerMixin):
             orig,
             mne_type,
             picks=picks,
-            verbose=False,
         )
